creation du service famillearticles

+++ src/main/java/com/skysoft/gestionstock/repository/FamilleArticlesRepository.java
@@ -3,5 +3,5 @@ package com.skysoft.gestionstock.repository;
 import com.skysoft.gestionstock.model.Famillearticles;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FamilleArticlesRepository extends JpaRepository<Famillearticles, Integer> {
+public interface FamilleArticlesRepository extends JpaRepository<Famillearticles, Long> {
 }
